pyproject.toml
[project]
name = "streamlit-chatbot-googleai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
]
